requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
numpy==1.21.0
oauthlib==3.1.1
pandas==1.2.5
pkg-resources==0.0.0
psycopg2==2.9.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
